Accident
Accidentally deleted my DrawingView, LaunchScreen.storyboard, and my Main.storyboard when I was trying to make DrawingView a member of the project so that the code would compile. Henricksen helped me to retrieve them.
Note to self: Next time I'll just ask him to help me make something a member of a project.